Fixed some size_t warnings
git-svn-id: http://openmesh.org/svnrepo/OpenMesh@849 fdac6126-5c0c-442c-9429-916003d36597

+++ trunk/src/Unittests/unittests_decimater.hh
@@ -49,14 +49,14 @@ TEST_F(OpenMeshDecimater, DecimateMesh) {
   HModQuadric hModQuadricDBG;
   decimaterDBG.add( hModQuadricDBG );
   decimaterDBG.initialize();
-  int removedVertices = 0;
+  size_t removedVertices = 0;
   removedVertices = decimaterDBG.decimate_to(5000);
                     decimaterDBG.mesh().garbage_collection();
 
-  EXPECT_EQ(2526, removedVertices) << "The number of remove vertices is not correct!";
+  EXPECT_EQ(2526, removedVertices)     << "The number of remove vertices is not correct!";
   EXPECT_EQ(5000u, mesh_.n_vertices()) << "The number of vertices after decimation is not correct!";
-  EXPECT_EQ(14994u, mesh_.n_edges()) << "The number of edges after decimation is not correct!";
-  EXPECT_EQ(9996u, mesh_.n_faces()) << "The number of faces after decimation is not correct!";
+  EXPECT_EQ(14994u, mesh_.n_edges())   << "The number of edges after decimation is not correct!";
+  EXPECT_EQ(9996u, mesh_.n_faces())    << "The number of faces after decimation is not correct!";
 }
 
 TEST_F(OpenMeshDecimater, DecimateMeshToFaceVerticesLimit) {
@@ -73,7 +73,7 @@ TEST_F(OpenMeshDecimater, DecimateMeshToFaceVerticesLimit) {
   HModQuadric hModQuadricDBG;
   decimaterDBG.add( hModQuadricDBG );
   decimaterDBG.initialize();
-  int removedVertices = 0;
+  size_t removedVertices = 0;
   removedVertices = decimaterDBG.decimate_to_faces(5000, 8000);
                     decimaterDBG.mesh().garbage_collection();
 
@@ -97,7 +97,7 @@ TEST_F(OpenMeshDecimater, DecimateMeshToFaceFaceLimit) {
   HModQuadric hModQuadricDBG;
   decimaterDBG.add( hModQuadricDBG );
   decimaterDBG.initialize();
-  int removedVertices = 0;
+  size_t removedVertices = 0;
   removedVertices = decimaterDBG.decimate_to_faces(4500, 9996);
                     decimaterDBG.mesh().garbage_collection();
